+++ src/pages/SignIn.js
@@ -1,28 +1,15 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { api } from "../../api/api";
-import Button from "../common/Button";
-import Input from "../common/Input";
+import { api } from "../api/api";
+import Button from "../components/common/Button";
+import Input from "../components/common/Input";
 
-function SignUp() {
+function SignIn() {
   const navigate = useNavigate();
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [isEmail, setIsEmail] = useState(false);
   const [isPassword, setIsPassword] = useState(false);
-
-  const handleClick = async () => {
-    try {
-      await api.post("/auth/signup", {
-        email,
-        password,
-      });
-
-      navigate("/signin");
-    } catch {
-      alert("동일한 이메일 있음");
-    }
-  };
 
   const handleEmail = (e) => {
     setEmail(e.target.value);
@@ -36,24 +23,35 @@ function SignUp() {
       setIsPassword(true);
     }
   };
-
   const handleButton = () => {
     if (isEmail && isPassword) {
       return false;
     } else return true;
   };
 
-  useEffect(() => {
-    //값 초기화
-    setIsEmail(false);
-    setIsPassword(false);
-  }, []);
-
+  const handleClick = async () => {
+    try {
+      const result = await api.post("/auth/signin", {
+        email,
+        password,
+      });
+      console.log(result);
+      alert("로그인 성공");
+      localStorage.setItem("token", result.data.access_token);
+      navigate("/todo");
+    } catch (error) {
+      if (error) {
+        alert("이메일이 없거나 비밀번호가 일치하지 않습니다.");
+      } else if (error.response.status === 404) {
+        alert("존재하지않는 이메일입니다.");
+      }
+    }
+  };
   return (
     <>
-      <header className="App-header">
-        <p>회원가입</p>
-      </header>
+      <div>
+        <h1>로그인</h1>
+      </div>
       <Input
         type="text"
         id="email"
@@ -72,13 +70,12 @@ function SignUp() {
       />
       <Button
         type="submit"
-        testid="signup-button"
+        testid="signin-button"
         onClick={handleClick}
         disabled={handleButton()}
-        text="회원가입"
+        text="로그인"
       />
     </>
   );
 }
-
-export default SignUp;
+export default SignIn;
